ChecksEmailIT: Move checker creation into the test methods
Not all tests need 2 required checkers. Creating the checkers where they
are used improves the readability of the tests. At the moment this may
not be a huge improvement, but I plan to add further tests that need
non-required checkers and I don't want to add further global checkers in
the test setup.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Icecadee7273632e607b00f8ea6b9a52a79925b30