Notify only owner if combined check state is updated to non-passing state
Reviewers are only interested in the combined check state when it turns
to a value that makes the change submittable. This is the case for the
states SUCCESSFUL and NOT_RELEVANT. Other combined check states are only
relevant for the change owner, hence for those an email notification
should only be sent to the change owner.

This is just the default for the notifications. Callers can always
overwrite it by setting NotifyHandling in the REST input.

Rerunning a check cannot cause an update of the combined check state to
SUCCESSFUL or NOT_RELEVANT, hence there is no test and documentation for
this case

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I37d407d3336f42d5cd876cccd90ee115cf7fe41d