Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I apologize for the chatty messages from Zuul.  I will work on fixing that before further testing.  You can filter them out by setting the "Comments only" switch.

This change and the issue 11730, linked to this change, is trying to solve the double notification problem: /1 from the CI, /2 from the checks plugin itself.

IOW Zuul should stop spamming the change comments, but only post the status outcome and the URL on the checks. The job of the checks plugin is to provide the right notification with direct links to the check jobs outcome in the combined mail notification (when this change is merged and the new version is deployed).

Patch-set: 1
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>