Update patch set 8
Patch Set 8:

Dhruv had asked me to look over the range of test and indicate whether they are sufficient. The test which was most important for me is "message is not shown if new check has no message", which now exists. Of course, we could have similar tests for all other properties as well. I trust your judgement whether those are necessary or not.

Patch-set: 8