Expose standalone rest api library
To access the plugin's pending checks REST endpoint from a standalone
CI system, expose standalone client library that shares the entities
classes from the api package. Isolate the entities classes in shared
library so that the classes are built only once. Extending the build
tool chain to allow to publish the library to Maven Central is outside
of the scope of this change and might be added in a follow-up change.

Change-Id: I536a6090ad88c9348b22075b367070be72cbfc9b