Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7: Code-Review-1
> > 
> > FluentLogger needs to be loaded by the user of the API. This shouldn't be necessary.
> 
> @DavidO: Is there a way to package FluentLogger (and other dependencies) as dependency into the jar published on maven, so that the user does not have to provide it manually?

Not really. The way it's working with Maven is: pom.xml should be exposed for each artifact, that will include all transitive dependencies. For this specific check client api library it would be google, gson, apache httpclient, google fluent-logger etc.

The consumer of this library would just provide Maven Coordinates for this artifact, and all transitive dependencies would be consumed from Maven Central as well (assumed that Maven build tool is used).

Also note, it's not sufficient, to just have fluent logger on the classpath, it must also be configured. Check init step in Gerrit itself, where a couple of fluent logger specific system properties are set.

Also note, that in follow-up change I haven't properly populated the dependencies section in pom.xml, mainly because it was quick and durty hack. Basically it's still a WIP series.

Patch-set: 7