Fire rerun-check event in rerun check endpoint
According to the plugin design document the trigger event must not be
explicitly tied to a change or patch set. Instead, it should rather be
a trigger for a CI system to call the pending checks endpoint again.

Rerun-check event can be processed by webhooks core plugin to propagate
this event to remote http endpoint. Webhook plugin is currently only
listening to the event derived from the project event.

Change-Id: I747947f644d3aae7b205c4fca5c755bc6790a5ea