Update patch set 1
Patch Set 1: Code-Review+1

I agree that it makes sense to align.

To improve this state, we would need to provide feedback to users on the fact that they are not on the latest patchset. The meaning is only conveyed by color and might not be straightforward to new users. There may be an opportunity to put "Go to latest patchset" closer to the Patchset navigation menu and only highlighting these 2 to show the relation between the button and the patchset navigation menu.

Maybe something we could discuss during on of the next UX office hours.

I'll LGTM this change and we can follow-up on improving it.

Patch-set: 1
Reviewer: Gerrit User 1079742 <1079742@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1