Perform migration from refs/meta/checkers/ to the same ref without slash
After the previous change corrected the name of the ref, this change has
to migrate the old ref to be the new ref.

The migration will be done whenever "GetCheck" or "CreateChecker" will
be called. We must do it before GetCheck since GetCheck is called many
times when using the checks API, and we must ensure all users perform
the migration.

In theory, users might create a checker before using the checks plugin
to get checks, and then it might cause an inconsistent state by having
two refs (since we will create a new checker in the new ref, while
still have the old ref). Therefore, before creating a checker we will
ensure the migration is already done so that we only ever have one ref
of that type.

Change-Id: Ic5c4e8188b9c6fb05258602063ff9617b44e8791