Perform migration from refs/meta/checkers/ to the same ref without slash
After the previous change corrected the name of the ref, this change has
to migrate the old ref to be the new ref.

The migration will be done upstream in the InitStep for the checks
plugin. When upgrading the plugin, Init should be run, and this is
documented in the init pgm notes.

Alternatively, we could use LifeCyclerListener, but then we would have
to continuously pay the price of opening the repository and checking for
the ref, so the current approach is cheaper.

Change-Id: Ic5c4e8188b9c6fb05258602063ff9617b44e8791