Add NoteDb representation of Override Check
In NoteDb, we will represent the overrides as a set of overrides. Each
element in the override will contain the id of the overrider, timestamp,
and reason for override.

This change adds a way to modify overrides of a specific check.
Error handling is not done here, and should be done in the REST endpoint
that adds the overrides.

Change-Id: I54fb67e2e26df5a7cb4090bdd30b0d68a3ef8226