Implement default OverridePolicy
In the previous change, OverridePolicy was created and it will be used
to determine whether or not a check impacts submission.

In this change, the default polcy is added: SingleOverridePolicy. It
returns an entity OverrideImpact. It checks if there is at least one
override, and if so, "overridden" will be set to true. Otherwise, it
will be set to false. Either way, the message is empty - it is supposed
to be used by other implementations of OverridePolicy.

Change-Id: I77cec7429b4e2f7f7113cb6a45eca93fe9c2bc81