Implement OverrideCheck
This change implements the logic of overriding checks. Also, it
implements the logic of calculating the submit impact in CheckInfo (if
ListChecksOption.SUBMIT_IMPACT is provided).

A follow-up change will have the calculation of submit impact in other
parts of the code that require that value to be accurate (checks submit
rule and combined check state).

A follow-up change will add all relevant documentation for the new
info/input objects.

Change-Id: I35986873aa56ab4176fe63b11d566226ef8185d9