Replace Set<BlockingCondition> with "required" boolean
Originally, BlockingCondition was implemented to allow more flexibility
with conditions that block submit.
However, we now see that this is unnecessary, and it's enough to have
just a boolean that says if this check blocks submission or not.

This change only adds the alternative to BlockingCondition for "Check",
and not for "Checker". As an optional follow-up change, it is possible
to do the same for "Checker", but might be more difficult due to NoteDb
migration.

Follow-up changes will have the "required" field entering another Info
entity for the purpose of having it easier to add more details once we
introduce "OverrideCheck".

Another follow-up change will have documentation updates.

This change only adds an alternative to BlockingCondition without
removing support for it. Thus, it doesn't break the frontend.
However, the frontend should still adjust by using the new fields. Once
the frontend is done, I can clean the backend by removing the legacy
BlockingCondition, as it is used only by the frontend.

Change-Id: I1790c62a6ee048fb7171b8a43ca95ca370dda210