Replace Set<BlockingCondition> with "required" boolean
Originally, BlockingCondition was implemented to allow more flexibility
with conditions that block submit.
However, we now see that this is unnecessary, and it's enough to have
just a boolean that says if this check blocks submission or not.

Follow-up changes will have the "required" field entering another Info
entity for the purpose of having it easier to add more details once we
introduce "OverrideCheck".

Another follow-up change will have documentation updates.

This change breaks the frontend. Frontend should adjust by using the new
field instead in all relevant Input/Info types. Best to wait until I
finish this chain, though.

Change-Id: I1790c62a6ee048fb7171b8a43ca95ca370dda210