Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> can you figure out why this wasn't caught by our *IT.java tests? JGit should be enforce naming conventions for refs, so something is broken in our use of JGit.

Are you sure JGit enforces naming conventions for refs? It seems like they don't.
E.g, in the JGit class "TestRepository" we use the method "branch" to create a ref, and there are very few checks there. Ideally, this JGit class is supposed to have their naming convention checks there (and in the same class that creates the refs themselves - not in a test class).

Alternatively, we could figure out all the refs that we have in Gerrit and write a script (or maybe somehow a test?) that runs "git check-ref-format" on them, to test for correct ref names.

Patch-set: 1