Allow for an online migration of checker ref
This commit is a follow-up to the checker ref migration. It allows
for an online migration, that is, a migration while the system is
running. To do so, we check if the legacy ref exists and use it,
if it does.
This way, we can roll out the new code without having to migrate
the refs in the same go. It also allows for rollbacks more easily.

The code will use the new ref, if the legacy ref does not exist.

We add a test to assert the behavior and apply some finishing
touches to the existing test code was well as move the migration
class from the 'api' to the 'db' package.

Change-Id: I86dea2bd1617cf0ec0db35b6cee59630e9e136a4