Update patch set 5
Patch Set 5:

> > Not completely. Thanks for untying the event from change and patch set. The event has to be even more general, though. The pending checks REST endpoint is on the host level. The event should be the same and hence not extend from ProjectEvent.
> 
> But event consumer could be interested in specific pojects only, and wouldn't trigger pending checks REST endpoint for other projects.
> 
> Moreover, there is currently a limitation of Webhooks plugin, that is only listening for ProjectEvent types as described in the commit message, see: [1]. TBH I don't see how it harms to provide project name in CheckTriggerEvent and derive that event from ProjectEvent.

I understand that some event consumers would like to filter on specific projects. It's definitely a use case we should eventually look into. By tying the CheckTriggerEvent to a ProjectEvent now, we might negatively impact other use cases or conflict with unimplemented features/improvements of checks, though. That's something I like to avoid. For instance, checkers might not be tied to a single repository in the future but multiple ones (or even all if we support something like *). Inheriting from ProjectEvent would mean that we'd need to send out many CheckTriggerEvents, which is not the planned behavior.

Hence, I would propose to go with a completely bare CheckTriggerEvent first. Any filtering possibilities is something I would like to add later when we also have a backup plan for still doing 'broadcast'/unfiltered like triggers when necessary.

For the Webhooks plugin, I'd suggest to always forward an event if it's not attached to a specific project.

Patch-set: 5
Reviewer: Gerrit User 1030912 <1030912@173816e5-2b9a-37c3-8a2e-48639d4f1153>