Fix broken CheckerRefMigratinTest
In [1] injection of AllProjects was replaced with an injection of
AllProjectsNameOnInitProvider since AllProjects was not available on
init.
The fact that the CheckerRefMigrationTest used that constructor was
missed which caused the tests to fail.

Keep the Project.Namekey of All-Projects as member and overload the
constructor so that tests can supply a test controlled Project.Namekey.

[1] 75e72bd - CheckerRefMigration: Use AllProjectsNameOnInitProvider

Bug: Issue 13005
Bug: Issue 12904
Change-Id: I1448c496b3613a07f7d51cba1b46d012c1506d3d