Remove validation of repository and name from CheckerConfig
Validation of repository and name was delegated to CheckerCreation and
CheckerUpdate in:

55931fe Validate checker properties in CheckerCreation and CheckerUpdate

Unfortunately the validation in CheckerConfig was kept together with a
loophole for tests to create CheckerCreations and CheckerUpdates
without validation so that the (outside of test-scope) unreachable
validation could be tested.

Together with validation remove #buildWithoutValidationForTesting()
from the builders of CheckerCreation and CheckerUpdate and the
test-cases in CheckerConfigTest that tested the unreachable validation.

Change-Id: I07bee31435700f92bc5749a1a27d4cc0aa8abf8e