Add support for new reboot ChecksApi
The plan is to switch the Checks Plugin to the new standard API and its
UI at some point. This will be done via an experiment flag.

This change adds basic support for this flag and converting/providing
its data to the new plugin API. At the moment the flag is turned off
for everyone, so this is currently for developing/debugging only.

Change-Id: Ib6b9f66f6bf34471db7b167495acccf7fac4d216