CheckSubmitRule: Return empty submit record if all checks are optional
If all configured checkers for a project are optional, it's better to
return an empty submit record instead of a record with status=OK.

This is also needed by submit requirements, since we convert results of
legacy submit rules to submit requirements. If the submit record is
empty, the SR evaluator will just skip creating a submit requirement
result for it.

Google-Bug_Id: b/209752314
Change-Id: Ia96861ed4bd5b4644f79988e3a7a8d221d42a227